break out what mdbook needs from the cli
