+++ crates/mdbook-goals/Cargo.toml
@@ -4,22 +4,11 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-anyhow = "1.0.86"
-disk-persist = "0.1.0"
-mdbook = "0.4.40"
-regex = "1.10.5"
-reqwest = { version = "0.12.5", features = ["blocking", "json"] }
-semver = "1.0.23"
-serde = "1.0.204"
-serde_json = "1.0.120"
-structopt = "0.3.26"
-walkdir = "2.5.0"
-rust_team_data = { git = "https://github.com/rust-lang/team" }
-lazy_static = "1.5.0"
-progress_bar = "1.0.5"
-chrono = "0.4.38"
-tokio = { version = "1.41.0", features = ["full"] }
-handlebars = { version = "6.1.0", features = ["dir_source"] }
-comrak = "0.29.0"
+anyhow = "1.0.94"
+mdbook = "0.4.43"
+regex = "1.11.1"
 rust-project-goals = { version = "0.1.0", path = "../rust-project-goals" }
 rust-project-goals-llm = { version = "0.1.0", path = "../rust-project-goals-llm" }
+semver = "1.0.23"
+serde_json = "1.0.133"
+structopt = "0.3.26"
