+++ crates/rust-project-goals-cli/Cargo.toml
@@ -4,3 +4,18 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+anyhow = "1.0.94"
+mdbook = "0.4.43"
+regex = "1.11.1"
+rust-project-goals = { version = "0.1.0", path = "../rust-project-goals" }
+rust-project-goals-llm = { version = "0.1.0", path = "../rust-project-goals-llm" }
+semver = "1.0.23"
+serde = "1.0.216"
+structopt = "0.3.26"
+walkdir = "2.5.0"
+tokio = { version = "1.41.0", features = ["full"] }
+serde_json = "1.0.133"
+chrono = "0.4.39"
+progress_bar = "1.0.6"
+handlebars = { version = "6.2.0", features = ["dir_source"] }
+comrak = "0.31.0"
