+++ crates/rust-project-goals-cli/src/generate_json.rs
