+++ crates/rust-project-goals-cli/src/templates.rs
