+++ crates/rust-project-goals/Cargo.toml
@@ -9,7 +9,7 @@ chrono = "0.4.39"
 disk-persist = "0.1.0"
 lazy_static = "1.5.0"
 regex = "1.11.1"
-reqwest = "0.12.9"
+reqwest = { version = "0.12.9", features = ["blocking", "json"] }
 serde = "1.0.216"
 serde_json = "1.0.133"
 walkdir = "2.5.0"
